Bug fix to Ldc.I4.S - was incorrectly treating int8 as byte not sbyte. Bug fix to Stelem - was not correctly loading StackSize when elements are ref-type due to failure to detect ref/value type difference properly.
Former-commit-id: 61f44bf01580acaa9521d050c0a9171daa493831

+++ Kernel/Compiler/Architectures/x86_32/Ldc.cs
@@ -81,7 +81,7 @@ namespace Kernel.Compiler.Architectures.x86_32
                     numBytes = 4;
                     break;
                 case OpCodes.Ldc_I4_S:
-                    iValue = (Int32)anILOpInfo.ValueBytes[0];
+                    iValue = (Int32)(sbyte)anILOpInfo.ValueBytes[0];
                     numBytes = 4;
                     break;
                 case OpCodes.Ldc_I8:
